backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
bcrypt==4.0.1
passlib[bcrypt]
pymongo
redis
python-jose[cryptography]
